+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Networking/Globals/GlobalDeclarations.cs
@@ -8,20 +8,23 @@
 
 global using System;
 global using System.Collections;
+global using System.Collections.Generic;
 global using System.ComponentModel;
-global using System.Diagnostics;
-global using System.Diagnostics.CodeAnalysis;
+global using System.Data;
+global using System.DirectoryServices;
 global using System.Drawing;
-global using System.Drawing.Drawing2D;
-global using System.Drawing.Imaging;
-global using System.Reflection;
+global using System.Globalization;
+global using System.IO;
+global using System.Net;
+global using System.Net.Sockets;
+global using System.Net.NetworkInformation;
 global using System.Runtime.InteropServices;
-global using System.Security.Principal;
 global using System.Security;
-global using System.Media;
 global using System.Text;
-global using System.Threading;
 global using System.Windows.Forms;
 
+global using Krypton.Toolkit.Suite.Extended.Common;
+
 global using Krypton.Toolkit.Suite.Extended.Developer.Utilities;
-global using Krypton.Toolkit.Suite.Extended.Common.Resources;
+
+global using Microsoft.WindowsAPICodePack.Dialogs;