+++ Documents/Help/Changelog.md
@@ -1,6 +1,7 @@
 =======
 
 ## 2022-04-04 - Build 2204 - April 2022
+* New `Krypton.Toolkit.Suite.Extended.Networking` module
 * Moved `KryptonPanelExtended` & `KryptonButtonPanel` to `Krypton.Toolkit.Suite.Extended.Panels` module
 * Added `Signed` configuration
 * Resolved [#268](https://github.com/Krypton-Suite/Extended-Toolkit/issues/268), Crashes using Outlook Grid control
@@ -15,7 +16,12 @@
 
 =======
 
-## 2022-02-03 - Build 2202 - February 2022
+## 2022-02-02 - Build 2202.1 - February 2022
+* Resolved [#299](https://github.com/Krypton-Suite/Extended-Toolkit/issues/299), Toolkit used is the wrong version
+
+=======
+
+## 2022-02-01 - Build 2202 - February 2022
 * Updated NuGet package information to aid deployment to GitHub
 * New `Krypton.Toolkit.Suite.Extended.Error.Reporting` module
 * Removal of unused 'using' statements